Only range-check signal_number if sizeof(long) > sizeof(int).
Use <limits.h> to determine if sizeof(long) > sizeof(int).

If it isn't, then comparing signal_number > INT_MAX for example is a condition that always yields FALSE, causing a compile-time warning in Watcom C.  Consequently, we need to use the pre-processor and not sizeof() to strip that part of code out to prevent generating the warning.